+++ src/lib.rs
@@ -87,6 +87,7 @@ pub mod vba;
 
 use serde::de::DeserializeOwned;
 use std::borrow::Cow;
+use std::cmp::{max, min};
 use std::fmt;
 use std::fs::File;
 use std::io::{BufReader, Read, Seek};
@@ -469,11 +470,18 @@ impl<T: CellType> Range<T> {
     pub fn get_value(&self, absolute_position: (u32, u32)) -> Option<&T> {
         let p = absolute_position;
         if p.0 >= self.start.0 && p.0 <= self.end.0 && p.1 >= self.start.1 && p.1 <= self.end.1 {
-            let idx = (absolute_position.0 - self.start.0) as usize * self.width()
-                + (absolute_position.1 - self.start.1) as usize;
-            return Some(&self.inner[idx]);
+            return self.get((
+                (absolute_position.0 - self.start.0) as usize,
+                (absolute_position.1 - self.start.1) as usize,
+            ));
         }
         None
+    }
+
+    /// Get cell value from **relative position**.
+    pub fn get(&self, relative_position: (usize, usize)) -> Option<&T> {
+        let (row, col) = relative_position;
+        self.inner.get(row * self.width() + col)
     }
 
     /// Get an iterator over inner rows
@@ -500,6 +508,14 @@ impl<T: CellType> Range<T> {
     /// Get an iterator over used cells only
     pub fn used_cells(&self) -> UsedCells<T> {
         UsedCells {
+            width: self.width(),
+            inner: self.inner.iter().enumerate(),
+        }
+    }
+
+    /// Get an iterator over all cells in this range
+    pub fn cells(&self) -> Cells<T> {
+        Cells {
             width: self.width(),
             inner: self.inner.iter().enumerate(),
         }
@@ -537,6 +553,88 @@ impl<T: CellType> Range<T> {
     {
         RangeDeserializerBuilder::new().from_range(self)
     }
+
+    /// Build a new `Range` out of this range
+    ///
+    /// # Remarks
+    ///
+    /// Cells within this range will be cloned, cells out of it will be set to Empty
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use calamine::{Range, DataType};
+    ///
+    /// fn example() {
+    ///     let mut a = Range::new((1, 1), (3, 3));
+    ///     a.set_value((1, 1), DataType::Bool(true));
+    ///     a.set_value((2, 2), DataType::Bool(true));
+    ///
+    ///     let b = a.range((2, 2), (5, 5));
+    ///     assert_eq!(b.get_value((2, 2)), Some(&DataType::Bool(true)));
+    ///     assert_eq!(b.get_value((3, 3)), Some(&DataType::Empty));
+    ///
+    ///     let c = a.range((0, 0), (2, 2));
+    ///     assert_eq!(c.get_value((0, 0)), Some(&DataType::Empty));
+    ///     assert_eq!(c.get_value((1, 1)), Some(&DataType::Bool(true)));
+    ///     assert_eq!(c.get_value((2, 2)), Some(&DataType::Bool(true)));
+    /// }
+    pub fn range(&self, start: (u32, u32), end: (u32, u32)) -> Range<T> {
+        let mut other = Range::new(start, end);
+        let (self_start_row, self_start_col) = self.start;
+        let (self_end_row, self_end_col) = self.end;
+        let (other_start_row, other_start_col) = other.start;
+        let (other_end_row, other_end_col) = other.end;
+
+        // copy data from self to other
+        let start_row = max(self_start_row, other_start_row);
+        let end_row = min(self_end_row, other_end_row);
+        let start_col = max(self_start_col, other_start_col);
+        let end_col = min(self_end_col, other_end_col);
+
+        if start_row > end_row || start_col > end_col {
+            return other;
+        }
+
+        let self_width = self.width();
+        let other_width = other.width();
+
+        // change referential
+        //
+        // we want to copy range: start_row..(end_row + 1)
+        // In self referencial it is (start_row - self_start_row)..(end_row + 1 - self_start_row)
+        let self_row_start = (start_row - self_start_row) as usize;
+        let self_row_end = (end_row + 1 - self_start_row) as usize;
+        let self_col_start = (start_col - self_start_col) as usize;
+        let self_col_end = (end_col + 1 - self_start_col) as usize;
+
+        let other_row_start = (start_row - other_start_row) as usize;
+        let other_row_end = (end_row + 1 - other_start_row) as usize;
+        let other_col_start = (start_col - other_start_col) as usize;
+        let other_col_end = (end_col + 1 - other_start_col) as usize;
+
+        {
+            let self_rows = self
+                .inner
+                .chunks(self_width)
+                .take(self_row_end)
+                .skip(self_row_start);
+
+            let other_rows = other
+                .inner
+                .chunks_mut(other_width)
+                .take(other_row_end)
+                .skip(other_row_start);
+
+            for (self_row, other_row) in self_rows.zip(other_rows) {
+                let self_cols = &self_row[self_col_start..self_col_end];
+                let other_cols = &mut other_row[other_col_start..other_col_end];
+                other_cols.clone_from_slice(self_cols);
+            }
+        }
+
+        other
+    }
 }
 
 impl<T: CellType> Index<usize> for Range<T> {
@@ -571,6 +669,39 @@ impl<T: CellType> IndexMut<(usize, usize)> for Range<T> {
     }
 }
 
+/// A struct to iterate over all cells
+#[derive(Debug)]
+pub struct Cells<'a, T: 'a + CellType> {
+    width: usize,
+    inner: ::std::iter::Enumerate<::std::slice::Iter<'a, T>>,
+}
+
+impl<'a, T: 'a + CellType> Iterator for Cells<'a, T> {
+    type Item = (usize, usize, &'a T);
+    fn next(&mut self) -> Option<Self::Item> {
+        self.inner.next().map(|(i, v)| {
+            let row = i / self.width;
+            let col = i % self.width;
+            (row, col, v)
+        })
+    }
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.inner.size_hint()
+    }
+}
+
+impl<'a, T: 'a + CellType> DoubleEndedIterator for Cells<'a, T> {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        self.inner.next_back().map(|(i, v)| {
+            let row = i / self.width;
+            let col = i % self.width;
+            (row, col, v)
+        })
+    }
+}
+
+impl<'a, T: 'a + CellType> ExactSizeIterator for Cells<'a, T> {}
+
 /// A struct to iterate over used cells
 #[derive(Debug)]
 pub struct UsedCells<'a, T: 'a + CellType> {
@@ -590,6 +721,23 @@ impl<'a, T: 'a + CellType> Iterator for UsedCells<'a, T> {
                 (row, col, v)
             })
     }
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let (_, up) = self.inner.size_hint();
+        (0, up)
+    }
+}
+
+impl<'a, T: 'a + CellType> DoubleEndedIterator for UsedCells<'a, T> {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        self.inner
+            .by_ref()
+            .rfind(|&(_, v)| v != &T::default())
+            .map(|(i, v)| {
+                let row = i / self.width;
+                let col = i % self.width;
+                (row, col, v)
+            })
+    }
 }
 
 /// An iterator to read `Range` struct row by row
@@ -603,4 +751,17 @@ impl<'a, T: 'a + CellType> Iterator for Rows<'a, T> {
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.as_mut().and_then(|c| c.next())
     }
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.inner
+            .as_ref()
+            .map_or((0, Some(0)), |ch| ch.size_hint())
+    }
 }
+
+impl<'a, T: 'a + CellType> DoubleEndedIterator for Rows<'a, T> {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        self.inner.as_mut().and_then(|c| c.next_back())
+    }
+}
+
+impl<'a, T: 'a + CellType> ExactSizeIterator for Rows<'a, T> {}